[render] move InputBindingsWidget
