+++ vclib/render/src/vclib/qt/gui/settings_dialog/input_bindings_widget.ui
